feat(server): expose `Accept` without httpX features (#2382)
